Add ReputationGuardTooltip on AddSpeakerForm and refactor it
