+++ app/components/Speakers/SpeakerPreview.jsx
@@ -13,7 +13,7 @@ import { ModalFormContainer } from '../Modal'
 import Icon from '../Utils/Icon'
 import ClickableIcon from '../Utils/ClickableIcon'
 import ReputationGuard from '../Utils/ReputationGuard'
-import { EditSpeakerForm } from './SpeakerForm'
+import EditSpeakerForm from './EditSpeakerForm'
 import ModalRemoveSpeaker from './ModalRemoveSpeaker'
 import { addModal } from '../../state/modals/reducer'
 import { removeSpeaker, updateSpeaker } from '../../state/video_debate/effects'
