--- app/components/Speakers/index.js
@@ -1,2 +1,0 @@
-export * from './SpeakerForm'
-export * from './SpeakerPreview'
