+++ app/styles/_components/Utils/reputation_guard_tooltip.sass
@@ -1,3 +1,6 @@
 .help-tooltip-trigger
-  display: inline-block
   cursor: help
+  display: inline-block
+
+  .Select-control
+    cursor: help
