+++ app/styles/_components/VideoDebate/video_debate.sass
@@ -43,6 +43,8 @@
   .actions
     padding: 15px 10px
     background: rgb(252, 252, 252)
+    .help-tooltip-trigger
+      display: block
 
 #video-show
   @media screen and (max-width: $widescreen)
