+++ app/styles/_global/icons.sass
@@ -4,8 +4,8 @@
 
 
 .fa
-  /* use !important to prevent issues with browser extensions that change fonts
   display: inline-block
+  /* use !important to prevent issues with browser extensions that change fonts */
   font-family: 'icomoon' !important
   speak: none
   font-size: inherit
@@ -19,7 +19,7 @@
   text-rendering: auto
   line-height: 1
 
-  /* Better Font Rendering ===========
+  /* Better Font Rendering */
   -webkit-font-smoothing: antialiased
   -moz-osx-font-smoothing: grayscale
 
@@ -42,3 +42,6 @@
 
 a.icon:not(:hover)
   color: $grey-dark
+
+a.icon[disabled]
+  color: $grey-lighter
