+++ app/styles/_libraries_overrides/react-select.sass
@@ -1,0 +1,2 @@
+.Select.is-disabled > .Select-control
+  background: lighten($grey-lighter, 10)
