+++ app/styles/application.sass
@@ -54,6 +54,7 @@
 @import "_helpers/is_blurred"
 @import "_helpers/quoted"
 @import "_libraries_overrides/react-popup"
+@import "_libraries_overrides/react-select"
 @import "_components/App/app"
 @import "_components/App/logo"
 @import "_components/App/sidebar"
